+++ veterinarian/src/Components/AreaDropdown/AreaDropDown.jsx
@@ -1,18 +1,27 @@
 import { useState } from 'react'
 import './areaDropdown.css'
-import neighborhood from './Area.js';
+import neighborhoods from './Area.js';
 function AreaDropdown() {
   const [area, setArea] = useState("")
+  const [filteredList, setFilteredList] = useState(neighborhoods);
+
+  const filterBySearch = (e) => {
+    const query = e.target.value;
+    let updatedList = [...neighborhoods];
+    updatedList = updatedList.filter((item) =>
+      item.toLocaleLowerCase().indexOf(query.toLocaleLowerCase()) !== -1)
+    setFilteredList(updatedList)
+  }
   return (
     <div className="dropdown">
       <div className="input-group mb-3">
-        <input type="text" className="s" aria-label="Text input with dropdown button" />
+        <input type="text" className="s" aria-label="Text input with dropdown button" onChange={filterBySearch} />
         <button dir='rtl' className="btn btn-outline-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">{area ? area : "סינון לפי שכונות"}</button>
         <ul className="dropdown-menu scrollable-menu">
           <li><a class="dropdown-item" dir='rtl' >בחר/י שכונה...</a></li>
           <li><hr class="dropdown-divider" /></li>
-          {neighborhood.map(e =>
-            <li><a className="dropdown-item" dir='rtl' onClick={() => setArea(e)}>{e}</a></li>
+          {filteredList.map((t, index) =>
+            <li key={index}><a className="dropdown-item" dir='rtl' onClick={() => setArea(t)}>{t}</a></li>
           )}
         </ul>
       </div>
